requirements.txt
appdirs==1.4.4
appnope==0.1.2
autopep8==1.5.4
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
distlib==0.3.1
docker==4.4.1
filelock==3.0.12
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.11.15
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
pycodestyle==2.6.0
Pygments==2.8.0
PyQt5==5.15.2
PyQt5-sip==4.19.24
python-dateutil==2.8.1
python-dotenv==0.15.0
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
TBB==0.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.4.0
virtualenv-clone==0.5.4
wcwidth==0.2.5
websocket-client==0.57.0
